requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
pillow==11.3.0
qrcode==8.2
python-decouple==3.8
